# 5.1 - Lazy router scopes