--- a/tests/TestCase/Routing/RouteBuilderTest.php
+++ b/tests/TestCase/Routing/RouteBuilderTest.php
@@ -715,7 +715,7 @@ class RouteBuilderTest extends TestCase
     public function testResourcesInScope($scopeMode): void
     {
         $builder = Router::createRouteBuilder('/');
-        $builder->useLazyScopes($scopeMode);
+        $builder->enableLazyScopes($scopeMode);
         $builder->scope('/api', ['prefix' => 'Api'], function (RouteBuilder $routes): void {
             $routes->setExtensions(['json']);
             $routes->resources('Articles');
@@ -916,7 +916,7 @@ class RouteBuilderTest extends TestCase
     public function testScope($scopeMode): void
     {
         $routes = new RouteBuilder($this->collection, '/api', ['prefix' => 'Api']);
-        $routes->useLazyScopes($scopeMode);
+        $routes->enableLazyScopes($scopeMode);
         $routes->scope('/v1', ['version' => 1], function (RouteBuilder $routes): void {
             $this->assertSame('/api/v1', $routes->path());
             $this->assertEquals(['prefix' => 'Api', 'version' => 1], $routes->params());
@@ -932,7 +932,7 @@ class RouteBuilderTest extends TestCase
     public function testScopeWithAction($scopeMode): void
     {
         $routes = new RouteBuilder($this->collection, '/api', ['prefix' => 'Api']);
-        $routes->useLazyScopes($scopeMode);
+        $routes->enableLazyScopes($scopeMode);
         $routes->scope('/prices', ['controller' => 'Prices', 'action' => 'view'], function (RouteBuilder $routes): void {
             $routes->connect('/shared', ['shared' => true]);
             $routes->get('/exclusive', ['exclusive' => true]);
@@ -1187,7 +1187,7 @@ class RouteBuilderTest extends TestCase
     public function testHttpMethodIntegration($scopeMode): void
     {
         $routes = new RouteBuilder($this->collection, '/');
-        $routes->useLazyScopes($scopeMode);
+        $routes->enableLazyScopes($scopeMode);
         $routes->scope('/', function (RouteBuilder $routes): void {
             $routes->get('/faq/{page}', ['controller' => 'Pages', 'action' => 'faq'], 'faq')
                 ->setPatterns(['page' => '[a-z0-9_]+'])
@@ -1243,7 +1243,7 @@ class RouteBuilderTest extends TestCase
     public function testUseLazyScopeInheritance(): void
     {
         $routes = new RouteBuilder($this->collection, '/');
-        $routes->useLazyScopes(true);
+        $routes->enableLazyScopes(true);
         $routes->scope('/admin', function ($builder) {
             $builder->connect('/home');
             $builder->scope('/pages', function ($builder) {