--- a/tests/TestCase/Routing/RouteCollectionTest.php
+++ b/tests/TestCase/Routing/RouteCollectionTest.php
@@ -832,7 +832,7 @@ class RouteCollectionTest extends TestCase
     public function testScopeResolution(): void
     {
         $routes = new RouteBuilder($this->collection, '/');
-        $routes->useLazyScopes(true);
+        $routes->enableLazyScopes(true);
         $routes->scope('/api', ['prefix' => 'Api'], function (RouteBuilder $builder) {
             $builder->connect('/version', 'Versions::current');
         });