This is an experiment that I had on recent travels. I spent a small amount of time with SPX profiling a simple 'hello world' template render. I wanted to see if there were any opportunities for quick wins. I found some interesting overhead in route generation. We've long known that route parsing can be a non-trivial amount of framework overhead.

I used routes from an application I have. It includes 65 routes organized into 10 scopes. I was interested in the performance of two functions. First, h `loadRoutes` method is responsible for loading application and plugin routes. Next, the `RoutingMiddleware::process` function is a great signal of the framework overhead, as it concludes with the controller being created and invoked. Currently in `5.x` we see:

`loadRoutes`

![before` loadroutes](https://github.com/cakephp/cakephp/assets/24086/653bacc4-c920-4cef-b37b-6f1400f6e6a1)

`RoutingMiddleware::process`
![before routingmiddleware](https://github.com/cakephp/cakephp/assets/24086/fb419654-94d1-44b5-a09e-f95f8cc98069)

My idea was that for applications that use routing scopes/prefixes/plugins we have a really simple way to defer route generation. Each of these concepts introduces a static prefix. When a request is processed we only need to load the scope that matches. The other scopes don't matter, and might not ever be needed. This could allow us to shave a measurable amount of time off the 'framework time'. 

I've implemented this theory and measured the inclusive time of `loadRoutes` and `RoutingMiddleware::process` again:

`loadRoutes`
![after loadroutes](https://github.com/cakephp/cakephp/assets/24086/e624274b-e47a-4e58-999d-5e33bb872e1c)

`RoutingMiddleware::process`
![after routingmiddleware](https://github.com/cakephp/cakephp/assets/24086/fc0f5c50-b4de-4afc-a53e-5e7c14fa678e)

By deferring route construction response times fell from 4ms to 3.2ms. The time spent in `process` went from 2.12ms to 1.6ms, which I think is decent savings.

I consider this a prototype and not yet production ready, but it 'works' and is measurably faster for the 'hello world' scenario. However, it creates some interesting tradeoffs. Those tradeoffs are:

1. While there are no 'breaks' in the API, there could be semantically different behavior in applications that access and manipulate routes directly. Applications that only use `Router` to generate URLs should not be impacted.
2. While we've optimized the simple scenario where a view generates 0 URLs, most responses *do* generate URLs. This means that when we go to generate URL strings we need to build **all** routes as `RouteCollection` doesn't have a good way to skip route scopes *yet*.
3. It is an overall increase incomplexity. This is another behavior that increases complexity with questionable value for real-world applications. Sure they may 'startup' faster, but we've not eliminated any work yet.

If we can find a reasonable solution to skipping scopes during URL matching we could deliver improved performance for more applications. One potential solution could be to use scope `params` as a way to fast-forward scopes that won't create routes that match anyways. This would work well with plugin based applications.

I also think this needs to be an opt-in change that applications enable. If it gets traction and doesn't generate many issues we can consider making it the default in the future.

So does this seem like it would be worth pursuing further?