@ADmad 2023-10-27
> This means that when we go to generate URL strings we need to build all routes as RouteCollection doesn't have a good way to skip route scopes yet.

So in the "real world" scenario this change won't have any impact on the total response time until we can make changes to URL generation too right?