@markstory 2023-10-27
> So in the "real world" scenario this change won't have any impact on the total response time until we can make changes to URL generation too right?

Yes. These changes would allow us to amortize the cost of building routes until later, but no actual savings for real world applications yet.

I will see how hacky parameter matching would be in `match()`. If that can be made to work we'd be able to provide savings to real world applications that use scopes + parameters.