@markstory 2023-10-30
I was able to implement partial scope resolution for scopes that use parameters. I think this gives some carrot to the way I think routes can be defined with simpler to read routes. Patterns like this

```php
    $builder->scope('/password', ['controller' => 'Users'], function ($builder) {
        $builder->connect('/reset', ['action' => 'resetPassword'], ['_name' => 'users:passwordReset']);
        $builder->connect('/new/{token}', ['action' => 'newPassword'], ['_name' => 'users:newPassword'])
            ->setPass(['token']);
    });
```

Provide performance improvements and make routes subjectively easier to read as you have layers of indentation in each of the major sections of your application. This also has benefits for applications composed of plugins and prefixes as these constructs include routing parameters for free. This could yield benefits for applications that put no effort into changing their routes and are already using prefixes or plugins with routes.