@markstory 2023-11-04
I've made lazy scopes an opt-in behavior now. I updated Router tests that use scopes to have run with both eager and lazy scope builders. Having this be an opt-in change should mitigate any breaking changes.

We could enable this by default in the `cakephp/app` for new applications, and ask application developers to give feedback on any performance gains.